+++ rpbp/run_all_rpbp_instances.py
@@ -18,7 +18,6 @@ logger = logging.getLogger(__name__)
 
 default_num_procs = 1
 default_tmp = None # utils.abspath('tmp')
-default_flexbar_format_option = None
 
 def main():
     parser = argparse.ArgumentParser(formatter_class=argparse.ArgumentDefaultsHelpFormatter,
@@ -31,9 +30,10 @@ def main():
 
     parser.add_argument('--tmp', help="The temp directory", default=default_tmp)
 
-    parser.add_argument('--flexbar-format-option', help="The name of the \"format\" "
-        "option for flexbar. This changed from \"format\" to \"qtrim-format\" in "
-        "version 2.7.", default=default_flexbar_format_option)
+    parser.add_argument('--flexbar-options', help="A space-delimited list of options to"
+        "pass to flexbar. Each option must be quoted separately and must include the"
+        "parameter value to be used, if required by flexbar. If specified, flexbar options"
+        "will override default settings.", nargs='*', type=str)
 
     parser.add_argument('--overwrite', help="If this flag is present, existing files "
         "will be overwritten.", action='store_true')
@@ -130,10 +130,10 @@ def main():
     if args.tmp is not None:
         tmp_str = "--tmp {}".format(args.tmp)
     
-    flexbar_format_option_str = ""
-    if args.flexbar_format_option is not None:
-        flexbar_format_option_str = "--flexbar-format-option {}".format(
-            args.flexbar_format_option)
+    flexbar_option_str = ""
+    if args.flexbar_options is not None:
+        flexbar_option_str = "--flexbar-options {}".format(' '.join('"' + flx_op + '"'
+            for flx_op in args.flexbar_options))
 
     
     # collect the job_ids in case we are using slurm and need to merge replicates
@@ -160,7 +160,7 @@ def main():
             logging_str, 
             star_str, 
             profiles_only_str,
-            flexbar_format_option_str, 
+            flexbar_option_str,
             keep_intermediate_str,
             mem_str
         )